Patched s3Client.destroy() in head-object.js

+++ nodes/head-object/head-object.js
@@ -97,7 +97,7 @@ function inputHandler(n, RED) {
       msg.key = payloadConfig.Key;
       send(msg);
     } finally {
-      s3Client.destroy();
+      if(s3Client) s3Client.destroy();
       /* Dereference vars */
       s3Client = null;
       /*********************/
